Extract validation of BP readings into a separate class
