+++ app/src/main/java/org/simple/clinic/bp/entry/BloodPressureEntrySheetEvents.kt
@@ -23,7 +23,7 @@ object BloodPressureRemoveClicked : UiEvent {
 
 object BloodPressureDiastolicBackspaceClicked : UiEvent
 
-data class BloodPressureBpValidated(val result: BloodPressureEntrySheetControllerV2.Validation) : UiEvent
+data class BloodPressureBpValidated(val result: BpValidator.Validation) : UiEvent
 
 data class BloodPressureDateValidated(val date: String) : UiEvent {
   fun result(validator: DateOfBirthFormatValidator) = validator.validate2(date)
